requirements.txt
antlr4-python3-runtime==4.9
matplotlib==3.3.2
networkx==2.6
